feat: add tagged quote dataset and local rotation
